backend/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
mangum>=0.17.0
supabase>=2.3.0    